chore: move dto into separate folders
