+++ src/blocks/NewsBlock/NewsBlock.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { Section } from '@/components/Section/Section';
-import { useNewsList } from '@/services/api/dto/News.query';
+import { useNewsList } from '@/services/api/dto/News/News.query';
 
 import { NewsCard } from './NewsCard';
 import { NewsCardSkeleton } from './NewsCardSkeleton';
