+++ src/blocks/ProductCollections/CollectionCard.tsx
@@ -2,7 +2,7 @@ import { ImageIcon } from 'lucide-react';
 import Image from 'next/image';
 import Link from 'next/link';
 
-import type { CollectionDTO } from '@/services/api/dto/Collection.dto';
+import type { CollectionDTO } from '@/services/api/dto/Collection/Collection.dto';
 
 type CollectionCardProps = Omit<CollectionDTO, 'id'> & {
   href: string;
