+++ src/blocks/ProductCollections/ProductCollections.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 
 import { Section } from '@/components/Section/Section';
-import { useCollections } from '@/services/api/dto/Collection.query';
+import { useCollections } from '@/services/api/dto/Collection/Collection.query';
 
 import { CollectionCard } from './CollectionCard';
 import { CollectionCardSkeleton } from './CollectionCardSkeleton';
