+++ src/blocks/ProductTabs/ProductTabs.tsx
@@ -14,7 +14,7 @@ import {
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
 import useHashParam from '@/hooks/useHashParam';
 import { cn } from '@/lib/utils';
-import type { ProductDetailsDTO } from '@/services/api/dto/Product.dto';
+import type { ProductDetailsDTO } from '@/services/api/dto/Product/Product.dto';
 
 const tabsData = [
   {
