+++ src/services/api/mockHandlers.ts
@@ -1,7 +1,7 @@
-import { collectionsHandlers } from './dto/Collection.mock';
-import { keyFeaturesHandlers } from './dto/KeyFeature.mock';
-import { newsHandlers } from './dto/News.mock';
-import { productHandlers } from './dto/Product.mock';
+import { collectionsHandlers } from './dto/Collection/Collection.mock';
+import { keyFeaturesHandlers } from './dto/KeyFeature/KeyFeature.mock';
+import { newsHandlers } from './dto/News/News.mock';
+import { productHandlers } from './dto/Product/Product.mock';
 
 export const mockHandlers = [
   ...productHandlers,
